+++ core/src/id_providers.rs
@@ -31,7 +31,7 @@ use crate::traits::IdProvider;
 use jsonrpsee_types::SubscriptionId;
 
 /// Generates random integers as subscription ID.
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct RandomIntegerIdProvider;
 
 impl IdProvider for RandomIntegerIdProvider {
@@ -42,7 +42,7 @@ impl IdProvider for RandomIntegerIdProvider {
 }
 
 /// Generates random strings of length `len` as subscription ID.
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct RandomStringIdProvider {
 	len: usize,
 }
@@ -62,7 +62,7 @@ impl IdProvider for RandomStringIdProvider {
 }
 
 /// No-op implementation to be used for servers that don't support subscriptions.
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct NoopIdProvider;
 
 impl IdProvider for NoopIdProvider {