+++ core/src/server/helpers.rs
@@ -28,7 +28,7 @@ use std::io;
 
 use crate::{to_json_raw_value, Error};
 use futures_channel::mpsc;
-use futures_util::stream::StreamExt;
+use futures_util::StreamExt;
 use jsonrpsee_types::error::{
 	CallError, ErrorCode, ErrorObject, ErrorResponse, CALL_EXECUTION_FAILED_CODE, OVERSIZED_RESPONSE_CODE,
 	OVERSIZED_RESPONSE_MSG, UNKNOWN_ERROR_CODE,